Translate minigame basics added, as well as a simple test for it (#79)
* Started work on translate minigame script

So far the script includes methods to setup a new swedish word's input field's on the screen, as well as a method to check if the inputs were correct.
TODO: highlight mistakes, trigger an event to tell the game that a word was correct

* Added way to test the translate minigame

Starting a new game works, setting up a new word works, and checking the word works as well

* Solved merge conflict

Why does Unity keep fucking with the font assets I will never know

* Started work on translate minigame script

So far the script includes methods to setup a new swedish word's input field's on the screen, as well as a method to check if the inputs were correct.
TODO: highlight mistakes, trigger an event to tell the game that a word was correct

* Added way to test the translate minigame

Starting a new game works, setting up a new word works, and checking the word works as well