Removed the old sound effect
I have been searching for this sound effect code for so long, at some point I thought that "oh maybe veera already removed it :)" NO it was still there but I hadn't tested out the translating minigame so I didn't know it was still there. Anyway the main problem with the sound effect playing when you're moving on to the next word is that it would overlap with the sound effect that plays when you press the button to move on.